refactor(frontend): change directory structure
