+++ apps/frontend/src/routes/_app/upload/.food-select.tsx
